[ml] feat: interfaces for different strategies
