+++ app/ml/strategies/project_ai.py
@@ -1,0 +1,10 @@
+from typing import List, Tuple
+
+from .abstract_strategy import AbstractAnalysisStrategy
+
+
+class ProjectAIStrategy(AbstractAnalysisStrategy):
+
+    async def analyze(self, pgn_data: str) -> List[Tuple[str, str]]:
+        # Implementation will go here
+        pass
